IETComm2020Hiwa_Sourcecode
This paper is published as a regular paper in IET Communication in the year 2020.

We develop a simulation environment using the Java programming language. We carry out simulation experiments on an Intel Core i5 2.4 GHz CPU and 4 GB RAM personal computer. In our simulator, we consider two Fog computing (FCs); each of them consists of a different number of Fog Nodes (FNs) and forwarding nodes and a data center with homogeneous Physical machines (PMs).